updated all components
Added a table to AllData, fixed logic on balance page to render appropriate text describing the balance/ transaction state, corrected the token value provided in CreateAccount to null instead of an empty string. added notification in Transaction.js to alert on successful transfer, corrected rendered text on Transaction.js to more accurately describe the current action Deposit/ Withdrawal, updated alert on Login.js to notify on failed login attempt, added ability to clear form and create a new user on CreateAccount.js, corrected the Navbar link to home from Link to NavLink, removed P tags for the signed in user natifacation and set NavLink.Text to display signed in user. added an alert to Home.js to greet the signed in user if one is signed in.