Updated styling and transaction logic
Styling improvements to some buttons as well as transaction display logic. A new button appears on the Deposit/ Withdrawal page offering to show all transactions instead of the last 7 if the transaction length exceeds 7. Added a navigation button for the home page in addition to the home page link on the NavBar brand element. Adjusted the card display on the AllData page.